pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "osint-iran-israel"
version = "0.1.0"
description = "Sistema de inteligencia multifuente — conflicto Irán-Israel-EE.UU."
requires-python = ">=3.11"

dependencies = [
    # Data acquisition
    "requests>=2.32",
    "feedparser>=6.0",          # RSS parsing (BBC, Al Jazeera, Google News)
    "atproto>=0.0.54",          # Bluesky API
    "opensky-api>=1.3",         # OpenSky flight data

    # Data processing
    "pandas>=2.2",
    "numpy>=1.26",
    "python-dotenv>=1.0",
    "pyarrow>=16",              # Parquet I/O

    # NLP & features
    "scikit-learn>=1.5",
    "nltk>=3.9",
    "sentence-transformers>=3.0",  # Embeddings

    # ML models
    # scikit-learn covers KNN, Naive Bayes, K-Means, Logistic/Linear Regression, Ridge, Lasso

    # Evaluation & analysis
    "matplotlib>=3.9",
    "seaborn>=0.13",
    "plotly>=5.22",

    # Notebook
    "jupyter>=1.1",
    "ipykernel>=6.29",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5",
    "black>=24",
    "pytest>=8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
